+++ dueca/GenericToken.cxx
@@ -77,12 +77,7 @@ const GlobalId& GenericToken::getChannelId() const
   return channel->getId();
 }
 
-bool GenericToken::localChannelEndExists(const NameSet& name)
-{
-  return ChannelManager::single()->findChannel(name) != NULL;
-}
-
-boost::weak_ptr<DCOMetaFunctor>
+std::weak_ptr<DCOMetaFunctor>
 GenericToken::getMetaFunctorBase(const std::string& fname) const
 {
   return DataClassRegistry::single().getMetaFunctor(dataclassname, fname);