+++ udpcom/PacketCommunicator.hxx
@@ -18,7 +18,7 @@
 #include <utility>
 
 #include <boost/intrusive_ptr.hpp>
-#include <boost/shared_ptr.hpp>
+#include <memory>
 
 #include <dueca_ns.h>
 #include <dueca/AsyncQueueMT.hxx>